Use uint64_t per Vitaly's feedback

+++ llvm/include/llvm/Transforms/Instrumentation/HWAddressSanitizer.h
@@ -67,7 +67,7 @@ enum { RuntimeMask = 0xffff };
 
 } // namespace HWASanAccessInfo
 
-std::optional<unsigned long long> getFixedShadowBase(void);
+std::optional<uint64_t> getFixedShadowBase(void);
 
 } // namespace llvm
